+++ src/manifest/parsers/nodeParser.ts
@@ -1,0 +1,22 @@
+import path = require("path");
+import { NodeMetaMap } from "../../domain";
+import { ManifestChangedHandler } from "../manifestChangedHandler";
+
+export class NodeParser {
+    static createModelMetaMap(nodesMap: any[]): Promise<NodeMetaMap> {
+        return new Promise((resolve) => {
+            const modelMetaMap: NodeMetaMap = new Map();
+            if (nodesMap === null || nodesMap === undefined) {
+                console.log("No nodes found in manifest!");
+                resolve(modelMetaMap);
+            }
+            Object.values(nodesMap)
+                .filter((model) => model.resource_type === ManifestChangedHandler.RESOURCE_TYPE_MODEL)
+                .forEach(({ name, root_path, original_file_path }) => {
+                    const fullPath = path.join(root_path, original_file_path);
+                    modelMetaMap.set(name, { path: fullPath });
+                });
+            resolve(modelMetaMap);
+        });
+    }
+}
